Move sponsored quota exceeded to Operator
